Update Dockerfile (#13)
Due to a bug with buildah, when trying to build a liberty project on che it fails due to not being able to parse sha values in a dockerfile.  This change replaces the sha value with the image value.   When buildah is fixed to accept sha values, we should change this back